refactor: fix convention issues in ported intelligence code
- Replace React.* namespace types with destructured imports
  (ReactNode, ChangeEvent, MouseEvent) in node-shell, library-view,
  avatar-management-modal
- Flatten constellation-view/hooks/ subdirectory — move hooks to
  component level matching codebase convention
