+++ apps/web/src/domains/intelligence/components/constellation-view/use-node-click-handlers.ts
